o minor changes -> less warnings (thanks to Georg & Georg) o 2 bugfixes in the interaction of fluid/adjoints/optimizer (still not working) o One bug resulted due to the functionality of the element's "LocationVector"-function:   There finally "NumDofPerNode" is used to evaluate the number of dofs based on the element data. If the discretization used for "LocationVector" is based on a different dofset (then usually one is scalar, the other vectorial), this results in a wrong number of dofs. Therefore i added a new function in the drt_utils which handles this case.
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@17005 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_opti/topopt_optimizer.cpp
@@ -486,6 +486,6 @@ Teuchos::RCP<DRT::ResultTest> TOPOPT::Optimizer::CreateFieldTest()
 
 /*----------------------------------------------------------------------*
  *----------------------------------------------------------------------*/
-const double TOPOPT::Optimizer::Iter() const {return optimizer_->Iter();}
+const int TOPOPT::Optimizer::Iter() const {return optimizer_->Iter();}
